requirements/base.txt
Django==4.2.3
django-environ==0.9.0
psycopg2==2.9.5
djangorestframework==3.14.0


whitenoise==6.5.0

django-filter==23.2
django-extensions==3.2.1
django-cors-headers==4.1.0
django-storages==1.13.1
django-simple-history==3.4.0
django-dynamic-fixture==4.0.1

drf-jwt==1.19.2

attrs==22.1.0

requests==2.31.0

pyotp==2.8.0
qrcode==7.4.2